user can login with auth0 but can't logout.
