+++ src/components/Account/LogoutAuth0.tsx
@@ -1,0 +1,11 @@
+import { useAuth0 } from "@auth0/auth0-react";
+
+const LogoutAuth0 = () => {
+    const {logout} = useAuth0();
+    
+    return (
+        {logout}
+    )
+}
+
+export default LogoutAuth0;
